gfxstream: mesa: import Mesa
Mesa has Linux WSI implementations that can be used for
gfxstream.  We have to first import the parts of Mesa that
we find useful.

Just modified

- guest/src/mesa/meson.build

to remove extra project args.

- guest/src/mesa/src/meson.build

to not automatically call subdir in all directories.

This also requires the VK_NO_NIR and Android build patches, which
are fairly upstreamable.  The strategy would be to merge all changes
required in the source code.

BUG=313466265
TEST=compile

Change-Id: I8f2240aa8c94fa976c64d97d1db24fbd64687234