One more fix for !usingDirectMapping.
The issue here is that with !usingDirectMapping
we end up copying over full pages which works
most of the time with GPU allocated memory.

The problem comes with some implementations
(like swiftshader) which essentually just use malloc.
This means we stomp over memory between allocations,
and can easily cause a crash.

BUG: 297588622
Test: Running a complex Vulkan application on swiftshader mac
Change-Id: Ic9cce9e82d3c45098de2a92bc62be1c43adea698