gfxstream: nuke the VK Stub Hal
For Cuttlefish/Trout/B*, when Vulkan isn't used (unlikely scenario),
the host launcher will not set "androidboot.hardware.vulkan=ranchu",
which will load the framework stub HAL.

Fuchsia guests always needs VK since GLES doesn't exist on the host.
The concept of stub is Android specific anyways.

For the Emulator, we have fc::setIfNotOverriden(fc::Vulkan, true)
in ./android-qemu2-glue/main.cpp, so I think it's always true with
the help of host Swiftshader.

The Emulator does have the ability to do "-feature -Vulkan" to stop
Vulkan support.  If it wants to retain that capability, a method
would be to modify boot properties not set
"androidboot.hardware.vulkan=ranchu" when Vulkan is explicitly
disabled.

The emulator seems to this ability in vl.c, with
"boot_property_add(..)".  Also, init.ranchu.rc will be have to
be modified to not set Vulkan true by default and let the host
do it.

Or we can just modify the host to not allow it to be overriden.
Whatever is the easiest.

For now, I just recommend not merging this to internal emulator
development branches.

BUG=298679844
TEST=launch_cvd

Change-Id: Idc1fb54f8c17640fb17cb725b173cfff7a361e10