Add a graphics detector
... which can be used to figure out what libs/gpus/drivers/quirks
are present on a machine.

This previously existed in Cuttlefish. However, due to
incompatibilities between Android built host tools and non-Android
built host GPU userspace drivers/libs, we would like to move this
to a prebuilt. Gfxstream is already available as a prebuilt and
the detector seems like it could be more broadly useful so let's
move it into Gfxstream.

Bug: b/297563172
Test: m gfxstream_graphics_detector && gfxstream_graphics_detector
Change-Id: I0e3267eaf948d41fc8f474c48c9d541b05173e05