Revert "gfxstream: nuke PortableMutex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39450015-6a0f-4206-93f8-ed4a16a57ec6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie7835c13213af25fd3883b96574af78faf52b3b3