Revert "gfxstream: nuke SocketStream"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39450015-6a0f-4206-93f8-ed4a16a57ec6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I687d4ea6f9fe9ec000581dec1897fafc669475c8