Revert "Update namespace in libandroidemu to gfxstream::guest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/012b359f-e99d-48b1-9b37-7c2e8b380e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f20cb79f9fd7c4a38a0b07cd653f167baf82f62