Revert "Update EGL to look for libs in same directory for host tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/012b359f-e99d-48b1-9b37-7c2e8b380e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic9df7e7632e2f993eb24c0589911f1dbcc3eeba5