Revert "Move existing vulkan codegen into nested vulkan-docs sub..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/012b359f-e99d-48b1-9b37-7c2e8b380e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5fc79b1c927e1cd0b839485752ab0a7a2083f4b9