Revert "Enable building guest Gfxstream components on host"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/012b359f-e99d-48b1-9b37-7c2e8b380e36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia3deda7b9d49e37665cb02f4f8b5f239c2a654bc