Add basic vulkan snapshot test
The test creates an instance, saves and loads it, to make sure it does
not crash.

In end2end snapshot tests, we do not have a real guest system. Thus
snapshot can only test the host part. To make it work we will need to
skip render thread and stream snapshots, and reuse all render channels
created before snapshots.

We also added APIs to ask render thread not to wait for snapshots, and
sync up with FrameBuffer after snapshot loads.

It runs with cuttlefish.

Bug: 294277842
Test: GfxstreamEnd2EndVkSnapshotBasicTest
Change-Id: Ibea5d9830f0e927e7dbe3552ed31ad901b5397ab