Add snasphot support for vk coherent memory
This commit tracks more functions related to coherent memory snapshot
and saves the memory content.

At this point we are not sure if we need to manually copy the memory
content, because they are supposed to be cloned in RAM snapshot. We
could revert that part if they turn out to be unnecessary in future.

Bug: 294277842
Test: GfxstreamEnd2EndVkSnapshotImageTest.HostMemoryContent
Change-Id: If5bf011beda6af83cf481c738614663475ffa15c