Add tests for coherent memory snapshot
This commit add tests for the coherent memory snapshot. The coherent
memory behaves significantly different in the end2end tests vs in AEMU,
thus the test does not provide full coverage.

Also, we need to modify some guest code because we were not able to
perserve the guest pointer after snapshot load in end2end tests.

Bug: 294277842
Test: GfxstreamEnd2EndVkSnapshotImageTest.HostMemoryContent
Change-Id: I0a8ecbfdbb3d1e8d7669be02e65addb4ac6b74ae