Update patch set 1
Patch Set 1:

We were unable to auto-submit your change due to the following error: Failed to submit 2 changes due to the following problems:

Change 2925047: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.

Change 2937981: Change could not be merged due to a path conflict. Please rebase the change locally and upload the rebased commit for review.



 view details: https://android-build.googleplex.com/presubmit-status?change_id=platform/hardware/google/gfxstream~2937981&revision_id=1&host=android&id=5728983390978048

Patch-set: 1
Tag: autogenerated:TreeHugger:Autosubmit-Fail
CC: Gerrit User 1974282 <1974282@85c56323-6fa9-3386-8a01-6480fb634889>