gfxstream: stream_renderer_snapshot(..) + stream_renderer_restore(..)
This is designed to match the API added in crrev.com/c/4853266.
The issue with the prior one is it assumes an android::base
C++ object is passed through, which would be difficult to get
working with crosvm or upstream QEMU.

Passing through a C-string to a directory is simpler and more
portable.

In addition, since the API is only used with the testing layer,
GFXSTREAM_ENABLE_HOST_SNAPSHOT is added.  This is because the
code assumes a newer AEMU, and the versioning on that is still
TBD.

BUG=324440526
TEST=compile

Change-Id: I71686603d8293781733fc5eb5fd936e46871baf7