Revert "[guest] Build fixes for Fuchsia"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8994564-59de-4afe-b905-134f3c73ed93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I84c3dd79efb60a6b3f1837b2ca453ad734d363c6