Handle dependency by VkMemoryDedicatedAllocateInfo
We have a situation that if an image memory is created with
VkMemoryDedicatedAllocateInfo, the following commands that refer to the
same image  must happen in this exact order:

vkCreateImage
vkAllocateMemory (with dedicated memory image)
vkBindImageMemory
vkCreateImageView

Our previous dependency graph implementation uses Vulkan objects as
nodes and does not haven enough granularity to handle this situation.
vkAllocateMemory will say the memory has dependency on the image.
vkBindImageMemory will say the image has dependency on the memory. Then
it creates a circular dependency.

In this commit we increase the granularity of the dependency graph to
use a pair <VkObjectHandle, State> as nodes. Currently State is an enum
with 2 values: CREATED and BOUND_MEMORY. By default, after a VkCreate*
command, an object will be in CREATED state. When calling
vkBindImageMemory or vkBindBufferMemory it will transform the image or
buffer into BOUND_MEMORY state. An image / buffer with BOUND_MEMORY will
have dependency of (1) itself being in CREATED state first (2) a memory
in CREATED state.

Then we can create a dependency graph that looks like this:

<VkImageView,CREATED> --> <VkImage,BOUND_MEMORY> --> <VkMemory,CREATED>
                                  |                           |
                                  |       ┌-------------------⅃
                                  v       v
                          <VkImage,CREATED>

No more circular dependency.

Test: GfxstreamEnd2EndVkSnapshotImageTests
Bug: 333447188
Change-Id: Ice45b04292eea151bed18988872e2f34401b7182