Revert "Allow static linking of libgfxstream_backend"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb6ededd-5bbb-4136-93e8-3e82fbd32b6f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If7aa3fe534130d34c6351e5c0bd8da6f9be1e3b9