Revert "Allow static linking of libgfxstream_backend"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd91dd50-17ba-435a-b3ce-eaf174e55238). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I55af5b5ee25e82f55373893df4ae6742d171efb9