gfxstream: Revert all unnecessary gfxstream_vk_* object types and corresponding translations in functable.
Original approach was to translate all possible handle types for complete
Mesa object equivalence. Practically, only a subset of these objects
actually need to be real Mesa objects for use with WSI. This simplifies
the translation and handwritten entries substantially, which is somewhat
brittle and bug-prone.

Clean up parts of functable.py that we don't need.

BUG=333133323
TEST=launch_cvd --gpu_mode=gfxstream_guest_angle

Change-Id: I106a45e45f4545d7255ddb63eeb82759e02a1284