Snapshot descriptor set contents
This commmit snapshots vulkan descriptor sets. Please refer to
go/aemu-vulkan-snapshot-testing for detailed design.

For descriptors, API playback does not work because it does not provide
enough granularity. To address this issue, we shadow the states of
descriptors and rebuild a descriptor update command on snapshot load.

Currently it only works with VulkanBatchedDescriptorSetUpdate.

Bug: 332770507
Test: CanSnapshotDescriptors
Change-Id: I01bd51cd03072da9357eba2404d408e6995ffc7a