Revert "Tweaking YCBCR conversion to return correct VK Format"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fac25ca9-6e89-44d1-88b4-b10bbcd9408b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6be31df5f359809c078b6687796f0cd94ea66072
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com