Add VK_EXT_debug_report extension support
Structures of this extension are used by many vulkan apps without
checking the feature support and this causes the emulator to crash
when an old system image is used.

Guest callback addresses within the structures added with this
extension are already being removed when creating an instance
on the host as this struct contains guest callback addresses.

Bug: b/332482903
Test: Run Khronos Vulkan samples
Change-Id: I474ad10ce36bc248d9e70210a7eee4d948197782