Revert "Remove binding structure from ANB image creation chain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f65ca156-af4b-40ba-8a84-ad2ad3257d3d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f552e98bc1fb843c49dfb2f8977d112f5383cce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com