Partial revert aosp/2982381 and aosp/3008938
... and add tests which reproduce the failures from BitmapTest
which must be fixed before re-enabling glProgramBinary support.

Prior to aosp/2982381 and aosp/3008938, the link status of
 program populated by glProgramBinary() would always return
unsuccessful because the host wasn't tracking it. With this,
it looks like skia would fallback to populating the program
by re-creating the shaders (via glCreateShader() + glShaderSource()
etc) and re-attaching the shaders.

After aosp/2982381 and aosp/3008938, it became possible to use
glGetProgramBinary() and glProgramBinary(). However:

Gfxstream emulates external images (AHBs in our case) using
normal textures on the host. Gfxstream also tracks information
about the uniforms/samplers/etc of shaders/programs and uses
this information to make decisions when emulating
GL_TEXTURE_EXTERNAL_OES as GL_TEXTURE_2D. Gfxstream also
rewrites samplerExternalOES to sampler2D in shaders to support
the emulation as the host is not using external images.

Overall, the issue seems to be that aosp/2982381 and aosp/3008938
is incomplete as glGetProgramBinary() returns effectively the binary
for the modified program (with samplerExternalOES converted to
sampler2D). With this, the guest does not know that the program
and its uniforms/samplers/etc are originally for usage with
GL_TEXTURE_EXTERNAL_OES and can not make the same emulation
decisions.

A future change could

 * Update glGetProgramBinary() to return a custom serialization
   of both the original-program/shaders and the binary of the
   modified-program/shaders

 * Update glProgramBinary() to deserialize the custom blob, use the
   original program/shaders for populating tracking information, and
   use the binary of the modified program/shaders to send to the
   host.

Bug: b/349821825
Test: GfxstreamEnd2EndTests
Change-Id: Ifa626c6d938e41ca37bb83fcad967942e751aa62