Revert "Fix Vulkan snapshot on MoltenVK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7b98c01-3c88-490b-9d6a-69a47bdebdd3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91d2c50b29b84472d6216547f140a402cedf579a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com