Revert "Fix Vulkan snapshot on MoltenVK"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff815d4d-9750-4012-ac6e-858ffca3b203). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6357f4ec64fce9bc84ad1c18b57ea70f614cfafd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com