gfxstream: guest: make sure signalSemaphoreValueCount is correct
From the Vulkan 1.3.204 spec:

VUID-VkSubmitInfo-pNext-03240

"If the pNext chain of this structure includes a
 VkTimelineSemaphoreSubmitInfo structure and any element of
 pSignalSemaphores was created with a VkSemaphoreType of
 VK_SEMAPHORE_TYPE_TIMELINE, then its signalSemaphoreValueCount
 member must equal signalSemaphoreCount"

Internally, Mesa WSI creates dummy semaphores/fences (see
transformVkSemaphore functions in in gfxstream_vk_private.cpp).
We don't want to forward that to the host, since there is no host
side Vulkan object associated with the dummy sync objects.

The way to test this behavior is Zink + glxgears, on Linux hosts.
It should fail without this check.

BUG=349122558
TEST=compile

Change-Id: I32e8856a1f98510904ee934125b5357ab217c644