Fix getDeadline() for large relative timeout values
Gfxstream uses async worker threads to wait on any sync fds when
vkWaitForFences() is called with any VkFence-s that are externally
backed by sync fds. The original calling thread than waits on a
`WaitGroup` which is effectively a condition variable + mutex +
atomic count of the number of sync fds still being waited upon.
The async worker threads decrement the atomic count whenever a
sync fd has been signaled and notifies the condition variable.

Prior to this change, a

  vkWaitForFences(... timeout = std::numeric_limits<uint64_t>::max() ...);

with an external fence that is not already signaled would result
in a call to `WaitGroup::waitAllLocked()` and then
`WaitGroup::conditionalTimeoutLocked()` with very large timeout
values. (Note: the 18446744073709551615 uint64_t max turns into
18446744073709551 as the `AndroidWorkPool` works in microseconds
but vkWaitForFences uses nanoseconds.) The
`conditionalTimeoutLocked()` function then handles effectively
the looping/waiting on the condition variable and checking if
the atomic count reaches 0.

When the condition is not yet satisfied, `conditionalTimeoutLocked()`
eventually calls into `pthread_cond_timedwait()` using a `timespec`
absolute time computed by `getDeadline()`.

Prior to this change, `getDeadline()` would compute the deadline
absolute time by adding the timeout microseconds to the current
"now" microseconds. It would then attempt to check for overflow
by comparing if the "now" is less than the addition. Note: with
"now" being `{ tv_sec: 1720822688 tv_usec: 834262 }` and a
`relative` timeout being
`18446744073709551`, the addition does not overflow and is
`18446744074543813`. Finally, `getDeadline()` would return this
newly computed value in the future back as `timespec` which
converts the microseconds to nanoseconds by multiplying by 1000.
However, this multiplication is not checked for overflow and
you can end up with a deadline that is in the past (see
playground in https://godbolt.org/z/1x5jnn811 ). If the deadline
ends up in the past, it seems like `pthread_cond_timedwait()`
may return immediatatly without giving up the mutex. This results
in the `conditionalTimeoutLocked()` function continuously looping
while holding the mutex and never allowing the worker thread to
grab the lock in `decrementBroadcast()`.

Bug: b/347288539
Test: cvd start --gpu_mode=gfxstream_guest_angle
Test: (x10) cts -m CtsGraphicsTestCases
Change-Id: Icdbc25c5f7afa756c02e3d87498b3d84ecc97c54