gfxstream: end2end: nuke RutabagaLayer, use kumquat for testing
This change nukes the RutabagaLayer and makes the
end2end testing layer use Kumquat.

The benefit is not having two different testing
frameworks.  Compared to the RutabagaLayer, Kumquat
is more complex, but potentially has more features
in the long run.  Some advanatages:

  - virtgpu_kumquat_ffi is portable enough that
    non-gfxstream libraries (minigbm, others) can
    depend on it.  This would enable more accurate
    gralloc testing long-term.

  - multiple different context types can connect to
    Kumquat at the same time, rather than just one with
    the RutabagaLayer.

  - If anyone is bored, it should be pretty easy add
    v4l2 apis via an virtio-media like interface.

  - The kumquat server can actually call
    stream_renderer_teardown(..) when performing
    snapshot tests.

  - Kumquat relies on EventFd + VkExternalSync:
    this can eliminate the need to export the
    libplatform sync API outside VK, EGL over the
    long-term.

For this change, the key thing is to connect to the right
Kumquat instance.  GoogleTest runs each test in a separate
thread, so we multiple kumquats are launched depending on
the number of tests launched.  The vkInitializeKumquat,
rcCreateDeviceKumquat and eglInitializeKumquat functions
were added to connect the right instance without exposing
gfxstream internal libraries.

A point of complexity was Gralloc.  We have an external
Gralloc instance, but HostConnection also maintains a
thread-local Gralloc instances.  The thread-local gralloc
instances can potentially step on the kumquat connection
of the pipe or ASG stream.  The solution was for
GrallocEmulated to maintain it's own virtgpu_kumquat
instance.

The proper long-term solution is move gralloc out of
HostConnection entirely.

BUG=354706346
TEST=Run end2end tests with and without change, observe
     same result:

[==========] 409 tests from 8 test suites ran. (433682 ms total)
[  PASSED  ] 381 tests.
[  SKIPPED ] 12 tests
[  FAILED  ] 16 tests

Change-Id: I0bcf1ba0fa2f18f2b30c4177681959b0815d4068