Revert "Reduce MoltenVK default logging level to errors"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bceb103-d085-44d5-9993-5c5b9cd9e70b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61282c0d0d213f55c0d130b7a53da1e5306518a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com