Introduce dirty bit for color buffer gl
We need to copy from gl to vk frequently. Let's skip the copying if
nothing changed since the last copy.

We did so by introducing a dirty bit for color buffer gl. We mark it as
dirty whenever gl contents are updated. Then Vulkan will only copy from
gl if the dirty bit is set.

Bug: 348710938
Change-Id: I19aa7ed34f9ccb4a08187a8b53297b55ad5ccd67