Revert "Remove unnecessary std::move"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/31b6cf1b-0226-4823-9df4-236ffb655629). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9874729d263d1ec76c2f60ab392fefd59cea856e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com