Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4fa6c0ea-2ab0-4915-859e-0a6475e265ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I56f8fb6b795c347d5ac37a28de1e5a5de26f724a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com