Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8efc607-7ef0-4563-9497-8e5e1bab545c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd34382b44aad3715505da1e8c4ad52eb88e86ce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com