Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6c71ec3-224d-4919-b008-545c2723047c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibf5df56de7c7fd03ce0b04ff2525d790da0613c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com