Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2a86c6e7-845a-4a0a-86c9-b8dc19012f9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9775a45eff6a0a5d179e3df60b9e583f240ca59a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com