Refactor VK destruction into locked-gather + unlocked-destroy
Upcoming changes will ensure that ANativeBufferInfo can not be
destroyed in the SyncThread to avoid use-after-free issues
(see b/360379616#comment3). To do this, image teardown needs to
be able to wait for other users (SyncThread) to finish. However,
the thread doing destruction should not hold the global lock
while waiting for the other users.

Bug: b/360379616
Test: Also cherry-pick aosp/3237181
Test: cvd start --gpu_mode=gfxstream_guest_angle_host_swiftshader
Test: atest com.android.cuttlefish.tests.FastbootRebootTest \
        --rerun-until-failure
Test: wait 30 minutes and see no failures
Change-Id: Ic3608c134429a747595948660374a883ce58e580