Revert "Bazel: Disable exceptions and link missing libs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04bc0f16-4067-4dea-b54d-75f7a53de529). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie32355e78080e84677ebf7055055e19a2d65c59f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com