Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f06c950e-8271-4abf-9a09-42e694ea7ae8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62e329ee64aab0835dba822a4dcfe45d26fc96a7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com