gfxstream: workaround presubmit SLO
The error is:

"Test run incomplete. Expected 416 tests, received 393
  Command run timed out after 900000 ms"

900000ms = 15 minutes, even the "native-test-timeout"
is 30mins here.  So perhaps a machine is buggy or
something.

By changing the order of the tests, a different test
will timeout in theory.  The has the affect of
extending the the SLO and confirming the issue is
with the timeout of the test-runner, not the test
itself.

BUG=362049750
TEST=CI

Change-Id: I8021af613ef601396eee0bb2620fccab337588a8