Revert "gfxstream: guest: don't rely on HostConnection for platf..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae48f71b-5b0f-427c-a936-5f27483aafb9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If56ca20d79b6fe6df6671565d1b2dde66b213dbd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com