Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0ead036c-e7cd-41d1-b604-bde0dfae4fb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07f270379c0f35c7ab2c40dd4f4ee1af9dc14fbf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com