Revert "gfxstream: guest: don't rely on HostConnection for platf..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d58049a5-30f2-4741-bf28-f05cd2b9021d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I03e266144fa93adcaacfabf205f6aa934d18bcce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com