Redirect custom gfxstream log to common logging
This sets up a redirect that allows the emulator to direct the logs from
the virtio-gpu-gfxstream-renderer macros to absl backend of the
emulator.

This will make sure we do not mix logging formats, which can cause
issues for android studio.

It also fixes (some unlikely) buffer overruns when the log message are
too long.

Bug: 368655950
Change-Id: If83de8f08e606ac53e88e94048d70844ec7a24b6