Revert "Update cb_handle_t for gralloc5"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3f5d6eb2-4ad9-41da-9453-97afab8cc03d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I692851656d5d5738ec5e4642f435eefa7c550847
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com