Revert "gfxstream: simplify GLESv1/GLESv2/EGL/Vulkan builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3dc1e136-eb66-4e51-b868-4b0ae7bc63b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia25d22569805b1b0f2468e1564625f22d876f75e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com